\name{plot.createArea}
\alias{plot.createArea}
\title{
plot method for an object from class of \code{createArea}
}
\description{
plot method for objects inherited from class of \code{createArea}
}
\usage{
\method{plot}{createArea}(x,...)
}
\arguments{
  \item{x}{an object from class of \code{createArea}}
  \item{...}{other objects that can be passed to \code{plot} function}
}
\author{
Rezgar Arabzadeh
}
\seealso{ \code{\link[WRSS]{createArea}} }
\keyword{ graphs }