\name{plot.sim}
\alias{plot.sim}
\title{
plot method for an WRSS object
}
\description{
plot method for objects inherited from class of \code{sim}
}
\usage{
\method{plot}{sim}(x,...)
}
\arguments{
  \item{x}{an object from class of \code{sim}}
  \item{...}{other objects that can be passed to \code{plot} function}
}
\author{
Rezgar Arabzadeh
}
\seealso{ \code{\link[WRSS]{sim}} }
\keyword{ graphs }