\docType{data}
\name{zarrineh}
\alias{zarrineh}
\title{data of Zarrineh-rud river basin}
\format{list object}
\description{
The \code{zarrineh} object, is a list of objects including time series and detail a five-reservoir systen in the Zarrineh-rud river basin.
}
\references{
Iran Water Resources Management Company, 2016.
}

