\name{WRSS-package}
\alias{WRSS-package}
\alias{WRSS}
\docType{package}
\title{
\packageTitle{WRSS}
}
\description{
\packageDescription{WRSS}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{WRSS}
\packageIndices{WRSS}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{WRSS}

Maintainer: \packageMaintainer{WRSS}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
