\docType{data}
\name{zarrineh}
\alias{zarrineh}
\title{data of Zarrineh-rud river basin}
\format{list object}
\source{
\url{https://doi.org/10.4211/hs.344b4f3b48e1476d91294fb3ee7fcb30}
}
\description{
The \code{zarrineh} object, is a list of objects including time series and detail a five-reservoir systen in the Zarrineh-rud river basin.
}
\references{
Iran Water Resources Management Company, 2016.
}

